Update NmHistorico type and connection string
The `NmHistorico` property in `HistoricoPallet.cs` was changed from `int` to `string` with a default value of `string.Empty`. A corresponding database migration was added to update the `nm_historico` column in the `historicopallet` table to `nvarchar(max)` and non-nullable. The `SiagDbContextModelSnapshot` was updated accordingly.

Additionally, the connection string for `SqlServerSIAGCrato` in `appsettings.json` was updated to change the server address from `192.168.0.6` to `192.168.0.8`.